# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TrustOMS Philippines API
  version: 1.0.0
  description: >
    Trust Order Management System — comprehensive API for trust banking operations
    including order lifecycle, trade execution, settlement, compliance, CRM,
    risk profiling, fee management, and client self-service portal.
  contact:
    name: TrustOMS Engineering
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: trustoms-access-token
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 25
    SearchParam:
      name: search
      in: query
      schema:
        type: string
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      expiresIn:
                        type: integer
        '400':
          description: Missing credentials
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange refresh token for new token pair
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
        '400':
          description: Missing refresh token
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders with filters
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: status
          in: query
          schema:
            type: string
        - name: portfolio_id
          in: query
          schema:
            type: string
        - name: side
          in: query
          schema:
            type: string
        - name: trader_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated order list
    post:
      tags: [Orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Order created
  /orders/auto-compute:
    post:
      tags: [Orders]
      summary: Auto-compute missing order field (quantity/price/amount)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Computed values
  /orders/pending-auth:
    get:
      tags: [Orders]
      summary: Pending authorization queue
      parameters:
        - name: tier
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Pending orders
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
        '404':
          description: Not found
    put:
      tags: [Orders]
      summary: Update draft order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Order updated
    delete:
      tags: [Orders]
      summary: Cancel order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
  /orders/{id}/submit:
    post:
      tags: [Orders]
      summary: Submit order for authorization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order submitted
  /orders/{id}/authorize:
    post:
      tags: [Orders]
      summary: Authorize order (requires senior/checker role)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approver_id, decision]
              properties:
                approver_id:
                  type: integer
                approver_role:
                  type: string
                decision:
                  type: string
                  enum: [APPROVED, REJECTED]
                comment:
                  type: string
      responses:
        '200':
          description: Authorization recorded
  /orders/{id}/reject:
    post:
      tags: [Orders]
      summary: Reject order (shortcut)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                approver_id:
                  type: integer
                comment:
                  type: string
      responses:
        '200':
          description: Order rejected
  /orders/{id}/revert:
    post:
      tags: [Orders]
      summary: Revert/un-cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order reverted
  /orders/{id}/timeline:
    get:
      tags: [Orders]
      summary: Order state transition history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Timeline events
  /orders/{id}/authorizations:
    get:
      tags: [Orders]
      summary: Get authorization chain for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authorization chain

  # ===========================================================================
  # Trades
  # ===========================================================================
...
  /trades/fills/order/{orderId}:
    get:
      tags: [Trades]
      summary: Get fills for a specific order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order fills

  # ===========================================================================
  # Confirmations
  # ===========================================================================
...